test(websocket): cover cookie auth in e2e suite; tidy docs wording
